+++ readme.md
@@ -299,25 +299,25 @@ ES Classes 표현방식을 알고, this를 통해서 객체를 다룰 수 있다
 - [x] 검색바 클릭
   - [x] 검색바 아래로 레이어가 뜬다
   - [x] 배경은 딤 처리 된다
-  - [] 추천 검색어 10개가 담긴다
-  - [] 키보드의 위, 아래 화살표를 통해 검색어 목록을 이동할 수 있다
-    - [] 이동하여 포커싱이 되면 배경색이 바뀐다
+  - [x] 추천 검색어 10개가 담긴다
+  - [x] 키보드의 위, 아래 화살표를 통해 검색어 목록을 이동할 수 있다
+    - [x] 이동하여 선택 되면 배경색이 바뀐다
 
-  - [] 검색한 내역이 있다면 
-    - [] 검색 목록이 5개까지 뜬다
-    - [] 추천 검색어 10개 위에 추가 된 형태
-    - [] 우측의 삭제 버튼이 있다
-      - [] 삭제 버튼을 누르면 검색어가 삭제된다(저장된 곳에서?)
-    - [] 이동하여 포커싱 되면 배경색이 바뀐다
+  - [x] 검색한 내역이 있다면 
+    - [x] 검색 목록이 5개까지 뜬다
+    - [x] 추천 검색어 10개 위에 추가 된 형태
+    - [x] 우측의 삭제 버튼이 있다
+      - [x] 삭제 버튼을 누르면 검색어가 삭제된다(저장된 곳에서?)
+    - [x] 이동하여 선택 되면 배경색이 바뀐다
 
-- [] 검색바 입력
-  - [] 검색바 아래로 레이어가 뜬다
-  - [] 딤 처리 
-  - [] 자동 완성 10개까지 표시
-    - [] 실제 이동은 없다
-  - [] 키보드의 위, 아래 화살표를 통해 자동 완성 목록을 이동할 수 있다
-    - [] 이동하여 포커싱이 되면 배경색이 바뀐다
-  - [] 데이터에 저장된다
+- [x] 검색바 입력
+  - [x] 검색바 아래로 레이어가 뜬다
+  - [x] 딤 처리 
+  - [x] 자동 완성 10개까지 표시
+    - [x] 실제 이동은 없다
+  - [x] 키보드의 위, 아래 화살표를 통해 자동 완성 목록을 이동할 수 있다
+    - [x] 이동하여 포커싱이 되면 배경색이 바뀐다
+  - [x] 데이터가 저장된다
