+++ src/js/events/search/api.js
@@ -1,8 +1,5 @@
-
-
-const URL = {
-  API: 'https://completion.amazon.com/api/2017/suggestions?session-id=133-4736477-7395454&customer-id=&request-id=4YM3EXKRH1QJB16MSJGT&page-type=Gateway&lop=en_US&site-variant=desktop&client-info=amazon-search-ui&mid=ATVPDKIKX0DER&alias=aps&b2b=0&fresh=0&ks=8&prefix=',
-};
+const URL =
+  'https://completion.amazon.com/api/2017/suggestions?session-id=133-4736477-7395454&customer-id=&request-id=4YM3EXKRH1QJB16MSJGT&page-type=Gateway&lop=en_US&site-variant=desktop&client-info=amazon-search-ui&mid=ATVPDKIKX0DER&alias=aps&b2b=0&fresh=0&ks=8&prefix=';
 
 export class DataFetcher {
   constructor(url) {
@@ -22,12 +19,11 @@ export class DataFetcher {
 
 export class APIClient extends DataFetcher {
   constructor(prefix) {
-    super(`${URL.API}${prefix}`);
+    super(`${URL}${prefix}`);
   }
 
   async getApiData() {
     const data = await this.getData();
-    console.log(data);
     return data.suggestions.map(el => el.value);
   }
 }
